web/requirements.txt
requests
numpy
Pillow
tornado
